requirements.txt
aiohttp==3.9.3
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0